17 accessing path parameters

+++ src/main/java/org/learning/javabrains/messenger/service/MessengerService.java
@@ -20,7 +20,7 @@ public class MessengerService {
 		return new ArrayList<Message>(messages.values());
 	}
 
-	public Message getMessage(int id) {
+	public Message getMessage(long id) {
 		Message message = messages.get(id);
 
 		return message;
